+++ mythtv/libs/libmythtv/mpeg/dvbdescriptors.h
@@ -674,7 +674,7 @@ class CountryAvailabilityDescriptor : public MPEGDescriptor
     QString CountryNames(void) const
     {
         QString countries="";
-        for (uint i=0; i<CountryCount(); i++)
+        for (size_t i=0; i<CountryCount(); i++)
         {
             if (i!=0) countries.append(" ");
             countries.append(QString::fromLatin1(
@@ -2652,7 +2652,7 @@ class FreesatLCNDescriptor : public MPEGDescriptor
 
         const unsigned char *payload = &data[2];
 
-        uint offset = 0;
+        size_t offset = 0;
         while ((offset + 5 < DescriptorLength()) &&
                (offset + 5 + payload[offset+4] <= DescriptorLength()))
         {
@@ -2677,19 +2677,19 @@ class FreesatLCNDescriptor : public MPEGDescriptor
     uint ServiceCount(void) const
         { return m_entries.size(); }
 
-    uint ServiceID(int i) const
+    uint ServiceID(size_t i) const
         { return *m_entries[i] << 8 | *(m_entries[i]+1); }
 
-    uint ChanID(int i) const
+    uint ChanID(size_t i) const
         { return (*(m_entries[i] + 2) << 8 | *(m_entries[i] + 3)) & 0x7FFF; }
 
-    uint LCNCount(int i) const
+    uint LCNCount(size_t i) const
         { return *(m_entries[i] + 4) / 4; }
 
-    uint LogicalChannelNumber(int i, int j) const
+    uint LogicalChannelNumber(size_t i, size_t j) const
         { return (*(m_entries[i] + 5 + j*4) << 8 | *(m_entries[i] + 5 + j*4 + 1)) & 0xFFF; }
 
-    uint RegionID(int i, int j) const
+    uint RegionID(size_t i, size_t j) const
         { return *(m_entries[i] + 5 + j*4 + 2) << 8 | *(m_entries[i] + 5 + j*4 + 3); }
 
     QString toString(void) const override; // MPEGDescriptor
@@ -2829,19 +2829,19 @@ class SkyLCNDescriptor : public MPEGDescriptor
     uint ServiceCount(void) const
         { return (DescriptorLength() - 2) / 9; }
 
-    uint ServiceID(int i) const
+    uint ServiceID(size_t i) const
         { return *(m_data + 4 + i*9) << 8 | *(m_data + 5 + i*9); }
 
-    uint ServiceType(int i) const
+    uint ServiceType(size_t i) const
         { return *(m_data + 6 + i*9); }
 
-    uint ChannelID(int i) const
+    uint ChannelID(size_t i) const
         { return *(m_data + 7 + i*9) << 8 | *(m_data + 8 + i*9); }
 
-    uint LogicalChannelNumber(int i) const
+    uint LogicalChannelNumber(size_t i) const
         { return *(m_data + 9 + i*9) << 8 | *(m_data + 10 + i*9); }
 
-    uint Flags(int i) const
+    uint Flags(size_t i) const
         { return *(m_data + 11 + i*9) << 8 | *(m_data + 12 + i*9); }
 
     QString toString(void) const override; // MPEGDescriptor
